backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
ecdsa==0.19.0
fastapi==0.114.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
jiter==0.5.0
openai==1.45.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psutil==7.0.0
psycopg==3.1.18
pyasn1==0.6.1
pydantic==2.9.1
pydantic_core==2.23.3
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.30.6
